backend/django/requirements.txt
Django
django-browser-reload
django-cors-headers
python-dotenv
PyJWT
pytest
pytest-django
